refactored update boundary
