+++ app/src/features/surveys/components/locations/SurveyAreaMapControl.tsx
@@ -5,14 +5,13 @@ import Box from '@mui/material/Box';
 import Button from '@mui/material/Button';
 import ComponentDialog from 'components/dialog/ComponentDialog';
 import FileUpload from 'components/file-upload/FileUpload';
-import { IUploadHandler } from 'components/file-upload/FileUploadItem';
 import MapContainer from 'components/map/MapContainer';
 import { ProjectSurveyAttachmentValidExtensions } from 'constants/attachments';
 import { FormikContextType } from 'formik';
-import { Feature, GeoJsonProperties, Geometry } from 'geojson';
+import { Feature } from 'geojson';
 import { LatLngBoundsExpression } from 'leaflet';
 import { useState } from 'react';
-import { handleGPXUpload, handleKMLUpload, handleShapeFileUpload } from 'utils/mapBoundaryUploadHelpers';
+import { boundaryUploadHelper } from 'utils/mapBoundaryUploadHelpers';
 import { ISurveyLocationForm } from '../StudyAreaForm';
 
 export interface ISurveyAreMapControlProps {
@@ -28,34 +27,6 @@ export const SurveyAreaMapControl = (props: ISurveyAreMapControlProps) => {
   const [updatedBounds] = useState<LatLngBoundsExpression | undefined>(undefined);
   const [isOpen, setIsOpen] = useState(false);
 
-  const boundaryUploadHandler = (): IUploadHandler => {
-    return async (file) => {
-      let features: Feature<Geometry, GeoJsonProperties>[] = [];
-      try {
-        if (file?.type.includes('zip') || file?.name.includes('.zip')) {
-          features = await handleShapeFileUpload(file);
-        } else if (file?.type.includes('gpx') || file?.name.includes('.gpx')) {
-          features = await handleGPXUpload(file);
-        } else if (file?.type.includes('kml') || file?.name.includes('.kml')) {
-          features = await handleKMLUpload(file);
-        }
-
-        // Map features into form data
-        const formData = features.map((item: Feature, index) => {
-          return {
-            name: `Study Area ${index + 1}`,
-            description: '',
-            geojson: [item],
-            revision_count: 0
-          };
-        });
-        setFieldValue(formik_key, formData);
-      } catch (error) {
-        setFieldError(formik_key, String(error));
-      }
-    };
-  };
-
   return (
     <>
       <ComponentDialog open={isOpen} dialogTitle="Import Boundary" onClose={() => setIsOpen(false)}>
@@ -64,7 +35,23 @@ export const SurveyAreaMapControl = (props: ISurveyAreMapControlProps) => {
             <Alert severity="info">If importing a shapefile, it must be configured with a valid projection.</Alert>
           </Box>
           <FileUpload
-            uploadHandler={boundaryUploadHandler()}
+            uploadHandler={boundaryUploadHelper({
+              onSuccess: (features: Feature[]) => {
+                // Map features into form data
+                const formData = features.map((item: Feature, index) => {
+                  return {
+                    name: `Study Area ${index + 1}`,
+                    description: '',
+                    geojson: [item],
+                    revision_count: 0
+                  };
+                });
+                setFieldValue(formik_key, formData);
+              },
+              onFailure: (message: string) => {
+                setFieldError(formik_key, message);
+              }
+            })}
             dropZoneProps={{
               acceptedFileExtensions: ProjectSurveyAttachmentValidExtensions.SPATIAL
             }}
